gnarkd/: added groth16.Verify api
